+++ docs/conf.py
@@ -151,9 +151,9 @@ copyright = '2018, MICO Authors'
 author = 'MICO Authors'
 
 # The short X.Y version
-version = '0.0.1'
+version = '0.3.0'
 # The full version, including alpha/beta/rc tags
-release = '0.0.1'
+release = '0.3.0'
 
 
 # -- General configuration ---------------------------------------------------