+++ src/edge.ts
@@ -111,8 +111,8 @@ export interface Edge {
     targetHandle?: LinkHandle;
     /** Edge type. Can be used for styling. */
     type?: any;
-    /** The id of the dynamic edge template to use for this edge. */
-    dynamicTemplate?: string;
+    /** The id of the path generator used for this edge. */
+    pathType?: string;
     /** List of markers to draw for this edge. */
     markers?: Marker[];
     /** Markers to draw at the start of this edge. */